applied new code style minor changes

+++ UnitTests/MugenMvvm.UnitTest/Extensions/BoxingExtensionsTest.cs
@@ -13,7 +13,7 @@ namespace MugenMvvm.UnitTest.Extensions
         [Fact]
         public void BoxBoolShouldUseCache()
         {
-            BoxingExtensions.Box((bool?)null).ShouldBeNull();
+            BoxingExtensions.Box((bool?) null).ShouldBeNull();
 
             var boxTrue = BoxingExtensions.Box(true);
             boxTrue.ShouldEqual(true);
@@ -23,11 +23,11 @@ namespace MugenMvvm.UnitTest.Extensions
             boxFalse.ShouldEqual(false);
             ReferenceEquals(BoxingExtensions.Box(false), boxFalse);
 
-            boxTrue = BoxingExtensions.Box((bool?)true);
+            boxTrue = BoxingExtensions.Box((bool?) true);
             boxTrue.ShouldEqual(true);
             ReferenceEquals(BoxingExtensions.Box(true), boxTrue);
 
-            boxFalse = BoxingExtensions.Box((bool?)false);
+            boxFalse = BoxingExtensions.Box((bool?) false);
             boxFalse.ShouldEqual(false);
             ReferenceEquals(BoxingExtensions.Box(false), boxFalse);
         }
@@ -35,7 +35,7 @@ namespace MugenMvvm.UnitTest.Extensions
         [Fact]
         public void BoxByteShouldUseCache()
         {
-            BoxingExtensions.Box((byte?)null).ShouldBeNull();
+            BoxingExtensions.Box((byte?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (byte i = 0; i < byte.MaxValue; i++)
             {
@@ -44,7 +44,7 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((byte?)i)!);
+                objects.Add(BoxingExtensions.Box((byte?) i)!);
             }
 
             objects.Count.ShouldEqual(byte.MaxValue);
@@ -53,7 +53,7 @@ namespace MugenMvvm.UnitTest.Extensions
         [Fact]
         public void BoxSByteShouldUseCache()
         {
-            BoxingExtensions.Box((sbyte?)null).ShouldBeNull();
+            BoxingExtensions.Box((sbyte?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (var i = sbyte.MinValue; i < sbyte.MaxValue; i++)
             {
@@ -62,7 +62,7 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((sbyte?)i)!);
+                objects.Add(BoxingExtensions.Box((sbyte?) i)!);
             }
 
             objects.Count.ShouldEqual(byte.MaxValue);
@@ -71,7 +71,7 @@ namespace MugenMvvm.UnitTest.Extensions
         [Fact]
         public void BoxUShortShouldUseCache()
         {
-            BoxingExtensions.Box((ushort?)null).ShouldBeNull();
+            BoxingExtensions.Box((ushort?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (ushort i = 0; i < BoxingExtensions.CacheSize; i++)
             {
@@ -80,21 +80,21 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((ushort?)i)!);
+                objects.Add(BoxingExtensions.Box((ushort?) i)!);
             }
 
             objects.Count.ShouldEqual(BoxingExtensions.CacheSize);
 
             objects.Clear();
-            objects.Add(BoxingExtensions.Box((ushort)(BoxingExtensions.CacheSize + 1)));
-            objects.Add(BoxingExtensions.Box((ushort)(BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((ushort) (BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((ushort) (BoxingExtensions.CacheSize + 1)));
             objects.Count.ShouldEqual(2);
         }
 
         [Fact]
         public void BoxShortShouldUseCache()
         {
-            BoxingExtensions.Box((short?)null).ShouldBeNull();
+            BoxingExtensions.Box((short?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (short i = -BoxingExtensions.CacheSize; i < BoxingExtensions.CacheSize; i++)
             {
@@ -103,21 +103,21 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((short?)i)!);
+                objects.Add(BoxingExtensions.Box((short?) i)!);
             }
 
             objects.Count.ShouldEqual(BoxingExtensions.CacheSize * 2);
 
             objects.Clear();
-            objects.Add(BoxingExtensions.Box((short)(BoxingExtensions.CacheSize + 1)));
-            objects.Add(BoxingExtensions.Box((short)(BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((short) (BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((short) (BoxingExtensions.CacheSize + 1)));
             objects.Count.ShouldEqual(2);
         }
 
         [Fact]
         public void BoxUIntShouldUseCache()
         {
-            BoxingExtensions.Box((uint?)null).ShouldBeNull();
+            BoxingExtensions.Box((uint?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (uint i = 0; i < BoxingExtensions.CacheSize; i++)
             {
@@ -126,21 +126,21 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((uint?)i)!);
+                objects.Add(BoxingExtensions.Box((uint?) i)!);
             }
 
             objects.Count.ShouldEqual(BoxingExtensions.CacheSize);
 
             objects.Clear();
-            objects.Add(BoxingExtensions.Box((uint)(BoxingExtensions.CacheSize + 1)));
-            objects.Add(BoxingExtensions.Box((uint)(BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((uint) (BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((uint) (BoxingExtensions.CacheSize + 1)));
             objects.Count.ShouldEqual(2);
         }
 
         [Fact]
         public void BoxIntShouldUseCache()
         {
-            BoxingExtensions.Box((int?)null).ShouldBeNull();
+            BoxingExtensions.Box((int?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (var i = -BoxingExtensions.CacheSize; i < BoxingExtensions.CacheSize; i++)
             {
@@ -149,21 +149,21 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((int?)i)!);
+                objects.Add(BoxingExtensions.Box((int?) i)!);
             }
 
             objects.Count.ShouldEqual(BoxingExtensions.CacheSize * 2);
 
             objects.Clear();
-            objects.Add(BoxingExtensions.Box((int)(BoxingExtensions.CacheSize + 1)));
-            objects.Add(BoxingExtensions.Box((int)(BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box(BoxingExtensions.CacheSize + 1));
+            objects.Add(BoxingExtensions.Box(BoxingExtensions.CacheSize + 1));
             objects.Count.ShouldEqual(2);
         }
 
         [Fact]
         public void BoxULongShouldUseCache()
         {
-            BoxingExtensions.Box((ulong?)null).ShouldBeNull();
+            BoxingExtensions.Box((ulong?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (ulong i = 0; i < BoxingExtensions.CacheSize; i++)
             {
@@ -172,21 +172,21 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((ulong?)i)!);
+                objects.Add(BoxingExtensions.Box((ulong?) i)!);
             }
 
             objects.Count.ShouldEqual(BoxingExtensions.CacheSize);
 
             objects.Clear();
-            objects.Add(BoxingExtensions.Box((ulong)(BoxingExtensions.CacheSize + 1)));
-            objects.Add(BoxingExtensions.Box((ulong)(BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((ulong) (BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((ulong) (BoxingExtensions.CacheSize + 1)));
             objects.Count.ShouldEqual(2);
         }
 
         [Fact]
         public void BoxLongShouldUseCache()
         {
-            BoxingExtensions.Box((long?)null).ShouldBeNull();
+            BoxingExtensions.Box((long?) null).ShouldBeNull();
             var objects = new HashSet<object>(ReferenceEqualityComparer.Instance);
             for (long i = -BoxingExtensions.CacheSize; i < BoxingExtensions.CacheSize; i++)
             {
@@ -195,14 +195,14 @@ namespace MugenMvvm.UnitTest.Extensions
 
                 objects.Add(item);
                 objects.Add(BoxingExtensions.Box(i));
-                objects.Add(BoxingExtensions.Box((long?)i)!);
+                objects.Add(BoxingExtensions.Box((long?) i)!);
             }
 
             objects.Count.ShouldEqual(BoxingExtensions.CacheSize * 2);
 
             objects.Clear();
-            objects.Add(BoxingExtensions.Box((long)(BoxingExtensions.CacheSize + 1)));
-            objects.Add(BoxingExtensions.Box((long)(BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((long) (BoxingExtensions.CacheSize + 1)));
+            objects.Add(BoxingExtensions.Box((long) (BoxingExtensions.CacheSize + 1)));
             objects.Count.ShouldEqual(2);
         }
